LayoutTests/platform/efl/fast/viewport/viewport-84-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument value "" for key "width" not recognized. Content ignored.
viewport size 64x480 scale 5.000000 with limits [5.000000, 5.000000] and userScalable -1.000000
